// packages/react-native-dogfood/gen/sfu_signal_rpc/signal.ts
export interface JoinRequest {
    /**
     * @generated from protobuf field: string subscriber_sdp_offer = 1;
     */
    subscriberSdpOffer: string;
    /**
     * @generated from protobuf field: string session_id = 3;
     */
    sessionId: string;
    /**
     * @generated from protobuf field: stream.video.sfu.CodecSettings codec_settings = 4;
     */
    codecSettings?: CodecSettings; // TODO: we should know if this is going to be
    // - publishing and subscribing, or just subscribing for future routing
}
// ...
class JoinRequest$Type extends MessageType<JoinRequest> {
    constructor() {
        super("stream.video.sfu.JoinRequest", [
            { no: 1, name: "subscriber_sdp_offer", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "session_id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "codec_settings", kind: "message", T: () => CodecSettings }
        ]);
    }
    create(value?: PartialMessage<JoinRequest>): JoinRequest {
        const message = { subscriberSdpOffer: "", sessionId: "" };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<JoinRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: JoinRequest): JoinRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string subscriber_sdp_offer */ 1:
                    message.subscriberSdpOffer = reader.string();
                    break;
                case /* string session_id */ 3:
                    message.sessionId = reader.string();
                    break;
                case /* stream.video.sfu.CodecSettings codec_settings */ 4:
                    message.codecSettings = CodecSettings.internalBinaryRead(reader, reader.uint32(), options, message.codecSettings);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: JoinRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string subscriber_sdp_offer = 1; */
        if (message.subscriberSdpOffer !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.subscriberSdpOffer);
        /* string session_id = 3; */
        if (message.sessionId !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.sessionId);
        /* stream.video.sfu.CodecSettings codec_settings = 4; */
        if (message.codecSettings)
            CodecSettings.internalBinaryWrite(message.codecSettings, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message stream.video.sfu.JoinRequest
 */
export const JoinRequest = new JoinRequest$Type();